+++ src/material/texture/texture.hpp
@@ -1,14 +1,14 @@
 #ifndef TEXTURE_H
 #define TEXTURE_H
 
-#include "vec.hpp"
+#include "../../geometry/vec.hpp"
 #include <memory>
 
-using vec3 = geometry::vec<double,3>;
+using vec3 = geometry::vec<float,3>;
 class texture
 {
   public:
-    virtual vec3 value(double u, double v, const vec3 &p) const = 0;
+    virtual vec3 value(float u, float v, const vec3 &p) const = 0;
 };
 
 class const_texture : public  texture
@@ -18,7 +18,7 @@ class const_texture : public  texture
 
     const_texture(){}
     const_texture(const vec3 c) :colour(c){}
-    virtual vec3 value(double u, double v, const vec3 &p) const
+    virtual vec3 value(float u, float v, const vec3 &p) const
     {
         return colour;
     }
@@ -38,9 +38,9 @@ class checker_texture : public texture
     checker_texture(const vec3 &v0, const vec3 &v1):
             even(std::make_shared<const_texture>(v0)),
             odd(std::make_shared<const_texture>(v1)) {}
-    virtual vec3 value(double u, double v, const vec3 &p) const
+    virtual vec3 value(float u, float v, const vec3 &p) const
     {
-        double sines = sin(10 * p[0])* sin(10* p[1])* sin(10 * p[2]);
+        float sines = sin(10 * p[0])* sin(10* p[1])* sin(10 * p[2]);
         if (sines <0)
         {
             return odd->value(u, v,p);
